refactor: extract MovieScreenig in its own file and rename
